requirements.txt
kafka-python==2.0.2
ray==2.8.0
pandas==2.1.3
pyarrow~=14.0.1
black==23.11.0